ref(auth): Break up authenticator code
Pull authenticator interface code out of Django models and into sentry.auth.authenticators.
