+++ src/sentry/api/endpoints/auth_index.py
@@ -102,7 +102,7 @@ class AuthIndexEndpoint(Endpoint):
         if "challenge" in validator.validated_data and "response" in validator.validated_data:
             try:
                 interface = Authenticator.objects.get_interface(request.user, "u2f")
-                if not interface.is_enrolled:
+                if not interface.is_enrolled():
                     raise LookupError()
 
                 challenge = json.loads(validator.validated_data["challenge"])
