+++ src/sentry/web/frontend/sudo.py
@@ -15,7 +15,7 @@ class SudoView(BaseSudoView):
 
         try:
             interface = Authenticator.objects.get_interface(request.user, "u2f")
-            if not interface.is_enrolled:
+            if not interface.is_enrolled():
                 raise LookupError()
         except LookupError:
             return False
