+++ tests/sentry/api/endpoints/test_user_authenticator_enroll.py
@@ -39,7 +39,7 @@ class UserAuthenticatorEnrollTest(APITestCase):
         assert email_log.info.call_args[1]["extra"]["message_type"] == email_type
 
     @mock.patch("sentry.utils.email.logger")
-    @mock.patch("sentry.models.TotpInterface.validate_otp", return_value=True)
+    @mock.patch("sentry.auth.authenticators.TotpInterface.validate_otp", return_value=True)
     def test_totp_can_enroll(self, validate_otp, email_log):
         # XXX: Pretend an unbound function exists.
         validate_otp.__func__ = None
@@ -49,7 +49,9 @@ class UserAuthenticatorEnrollTest(APITestCase):
             kwargs={"user_id": "me", "interface_id": "totp"},
         )
 
-        with mock.patch("sentry.models.authenticator.generate_secret_key", return_value="Z" * 32):
+        with mock.patch(
+            "sentry.auth.authenticators.base.generate_secret_key", return_value="Z" * 32
+        ):
             resp = self.client.get(url)
 
         assert resp.status_code == 200
@@ -70,7 +72,7 @@ class UserAuthenticatorEnrollTest(APITestCase):
 
         # also enrolls in recovery codes
         recovery = Authenticator.objects.get_interface(user=self.user, interface_id="recovery")
-        assert recovery.is_enrolled
+        assert recovery.is_enrolled()
 
         self._assert_security_email_sent("mfa-added", email_log)
 
@@ -81,7 +83,7 @@ class UserAuthenticatorEnrollTest(APITestCase):
         assert resp.status_code == 400
 
     @mock.patch("sentry.utils.email.logger")
-    @mock.patch("sentry.models.TotpInterface.validate_otp", return_value=False)
+    @mock.patch("sentry.auth.authenticators.TotpInterface.validate_otp", return_value=False)
     def test_invalid_otp(self, validate_otp, email_log):
         # XXX: Pretend an unbound function exists.
         validate_otp.__func__ = None
@@ -99,8 +101,8 @@ class UserAuthenticatorEnrollTest(APITestCase):
         assert email_log.call_count == 0
 
     @mock.patch("sentry.utils.email.logger")
-    @mock.patch("sentry.models.SmsInterface.validate_otp", return_value=True)
-    @mock.patch("sentry.models.SmsInterface.send_text", return_value=True)
+    @mock.patch("sentry.auth.authenticators.SmsInterface.validate_otp", return_value=True)
+    @mock.patch("sentry.auth.authenticators.SmsInterface.send_text", return_value=True)
     def test_sms_can_enroll(self, send_text, validate_otp, email_log):
         # XXX: Pretend an unbound function exists.
         validate_otp.__func__ = None
@@ -152,7 +154,7 @@ class UserAuthenticatorEnrollTest(APITestCase):
             assert resp.status_code == 400
 
     @mock.patch("sentry.utils.email.logger")
-    @mock.patch("sentry.models.U2fInterface.try_enroll", return_value=True)
+    @mock.patch("sentry.auth.authenticators.U2fInterface.try_enroll", return_value=True)
     def test_u2f_can_enroll(self, try_enroll, email_log):
         new_options = settings.SENTRY_OPTIONS.copy()
         new_options["system.url-prefix"] = "https://testserver"
@@ -264,15 +266,15 @@ class AcceptOrganizationInviteTest(APITestCase):
         assert om.user is None
         assert om.email == "newuser@example.com"
 
-    @mock.patch("sentry.models.U2fInterface.try_enroll", return_value=True)
+    @mock.patch("sentry.auth.authenticators.U2fInterface.try_enroll", return_value=True)
     def test_accept_pending_invite__u2f_enroll(self, try_enroll):
         om = self.get_om_and_init_invite()
         resp = self.setup_u2f()
 
         self.assert_invite_accepted(resp, om.id)
 
-    @mock.patch("sentry.models.SmsInterface.validate_otp", return_value=True)
-    @mock.patch("sentry.models.SmsInterface.send_text", return_value=True)
+    @mock.patch("sentry.auth.authenticators.SmsInterface.validate_otp", return_value=True)
+    @mock.patch("sentry.auth.authenticators.SmsInterface.send_text", return_value=True)
     def test_accept_pending_invite__sms_enroll(self, send_text, validate_otp):
         # XXX: Pretend an unbound function exists.
         validate_otp.__func__ = None
@@ -310,7 +312,7 @@ class AcceptOrganizationInviteTest(APITestCase):
 
         self.assert_invite_accepted(resp, om.id)
 
-    @mock.patch("sentry.models.TotpInterface.validate_otp", return_value=True)
+    @mock.patch("sentry.auth.authenticators.TotpInterface.validate_otp", return_value=True)
     def test_accept_pending_invite__totp_enroll(self, validate_otp):
         # XXX: Pretend an unbound function exists.
         validate_otp.__func__ = None
@@ -337,7 +339,7 @@ class AcceptOrganizationInviteTest(APITestCase):
         self.assert_invite_accepted(resp, om.id)
 
     @mock.patch("sentry.api.endpoints.user_authenticator_enroll.logger")
-    @mock.patch("sentry.models.U2fInterface.try_enroll", return_value=True)
+    @mock.patch("sentry.auth.authenticators.U2fInterface.try_enroll", return_value=True)
     def test_user_already_org_member(self, try_enroll, log):
         om = self.get_om_and_init_invite()
         self.create_existing_om()
@@ -351,7 +353,7 @@ class AcceptOrganizationInviteTest(APITestCase):
         )
 
     @mock.patch("sentry.api.endpoints.user_authenticator_enroll.logger")
-    @mock.patch("sentry.models.U2fInterface.try_enroll", return_value=True)
+    @mock.patch("sentry.auth.authenticators.U2fInterface.try_enroll", return_value=True)
     def test_org_member_does_not_exist(self, try_enroll, log):
         om = self.get_om_and_init_invite()
 
@@ -369,7 +371,7 @@ class AcceptOrganizationInviteTest(APITestCase):
         assert log.error.call_args[0][0] == "Invalid pending invite cookie"
 
     @mock.patch("sentry.api.endpoints.user_authenticator_enroll.logger")
-    @mock.patch("sentry.models.U2fInterface.try_enroll", return_value=True)
+    @mock.patch("sentry.auth.authenticators.U2fInterface.try_enroll", return_value=True)
     def test_invalid_token(self, try_enroll, log):
         om = self.get_om_and_init_invite()
 
@@ -384,7 +386,7 @@ class AcceptOrganizationInviteTest(APITestCase):
         assert om.email == "newuser@example.com"
 
     @mock.patch("sentry.api.endpoints.user_authenticator_enroll.logger")
-    @mock.patch("sentry.models.U2fInterface.try_enroll", return_value=True)
+    @mock.patch("sentry.auth.authenticators.U2fInterface.try_enroll", return_value=True)
     def test_enroll_without_pending_invite__no_error(self, try_enroll, log):
         new_options = settings.SENTRY_OPTIONS.copy()
         new_options["system.url-prefix"] = "https://testserver"
