+++ tests/sentry/models/test_organization.py
@@ -3,6 +3,7 @@ from __future__ import absolute_import
 import copy
 from sentry.utils.compat import mock
 
+from sentry.auth.authenticators import TotpInterface
 from sentry.models import (
     ApiKey,
     AuditLogEntry,
@@ -21,7 +22,6 @@ from sentry.models import (
     ReleaseEnvironment,
     ReleaseFile,
     Team,
-    TotpInterface,
     User,
 )
 from sentry.testutils import TestCase
